Basic network sync example -- only sync node list (no element data)
This uses an in-memory backing for PouchDB in the browser, which means that it doesn't work for offline support.  We'd need to add another PouchDB instance that is backed by an in-browser DB (like IndexDB).

We can set up one-way replication from the in-memory DB to the browser DB.  It could be selectively disabled for changes that aren't made locally.